added 'init' file with dynamic package import
This init file dynamically imports modules from this directory.
It checks to make sure that the imported module contains
a Class with a SubClass of type SwitcherBaseClass, the abstract
base class provided in the switcher package.

All this init file does is save the user the trouble of having to modify
this init file every time a new switch module is added, replacing:
"""
from .switch import Switch as switch
"""