added switch method to SwitcherBaseClass
The purpose of the switch method is to exactly match the string
passed to the 'argument' parameter to the name of a function
that exists in the class that uses the SwitchBaseClass as its subclass.

Since the switch method is always the same syntax, no matter
what class calls it, I replaced the original abstract switch function
with this one.