added 'test_month_init' pytest function
This function tests the 'Month' class in the 'month' module
by calling the 'numbers_to_months' function with a range from
zero to the length of the 'months' list.

This tests proves that:
- The sample switch provided converts an integer to the corresponding
month.
- The SwitcherBaseClass switch function successfully curries
a function name from the input
- The init module in the 'switches' package successfully dynamically
imports the modules that contain classes of the 'SwitcherBaseClass'
subtype
- The 'numbers_to_months' function successfully talks to the underlying 'switch' function from the 'SwitcherBaseClass'
- Passing an invalid argument to the underlying 'switch' function
return the message 'Invalid request'